+++ test/comparison-tests/babel-issue81/webpack.config.js
@@ -1,0 +1,18 @@
+module.exports = {
+    entry: './a.ts',
+    output: {
+        filename: 'bundle.js'
+    },
+    resolve: {
+        extensions: ['', '.ts', '.js']
+    },
+    devtool: 'source-map',
+    module: {
+        loaders: [
+            { test: /\.ts$/, loader: 'babel-loader?presets[]=es2015!ts-loader' }
+        ]
+    }
+}
+
+// for test harness purposes only, you would not need this in a normal project
+module.exports.resolveLoader = { alias: { 'ts-loader': require('path').join(__dirname, "../../index.js") } }