+++ test/execution-tests/1.8.2_babel-allowSyntheticDefaultImports/src/simple.ts
@@ -1,0 +1,5 @@
+import mathutils from "mathutils";
+
+export function usingMathutils() {
+    return `with ${ mathutils.version } we can make ${ mathutils.adder(3, 4) }`;
+}