+++ test/execution-tests/1.8.2_babel-allowSyntheticDefaultImports/test/simple.tests.ts
@@ -1,0 +1,7 @@
+import { usingMathutils } from "../src/simple";
+
+describe("simple", () => {
+  it("usingMathutils successfully uses the mathutils module", () => {
+    expect(usingMathutils()).toBe("with 1.2 we can make 7");
+  });
+});