+++ test/execution-tests/2.0.3_typesResolution/karma.conf.js
@@ -1,0 +1,55 @@
+/* eslint-disable no-var, strict */
+'use strict';
+var path = require('path');
+var webpack = require('webpack');
+var webpackConfig = require('./webpack.config.js');
+
+module.exports = function(config) {
+  config.set({
+    browsers: [ 'PhantomJS' ],
+
+    files: [
+      // This loads all the tests
+      'main.js'
+    ],
+
+    port: 9876,
+
+    frameworks: [ 'jasmine' ],
+
+    logLevel: config.LOG_INFO, //config.LOG_DEBUG
+
+    preprocessors: {
+      'main.js': [ 'webpack', 'sourcemap' ]
+    },
+
+    webpack: {
+      devtool: 'inline-source-map',
+      debug: true,
+      module: {
+          loaders: webpackConfig.module.loaders
+      },
+      resolve: webpackConfig.resolve,
+
+      // for test harness purposes only, you would not need this in a normal project
+      resolveLoader: webpackConfig.resolveLoader
+    },
+
+    webpackMiddleware: {
+      quiet: true,
+      stats: {
+        colors: true
+      }
+    },
+
+    // reporter options
+    mochaReporter: {
+      colors: {
+        success: 'bgGreen',
+        info: 'cyan',
+        warning: 'bgBlue',
+        error: 'bgRed'
+      }
+    }
+  });
+};