+++ test/execution-tests/babel-es2015/src/simple.ts
@@ -1,0 +1,7 @@
+export function adder(num1: number, num2: number) {
+    return num1 + num2;
+}
+
+const myValue = 'Edmund';
+
+export default myValue;