+++ test/execution-tests/babel-es6resolveParent/src/submodule/index.tsx
@@ -1,0 +1,5 @@
+import { BaseComponent } from '../index';
+
+export class Component extends BaseComponent {}
+
+export default 'foo';