+++ test/execution-tests/simpleDependency/src/deeperDep.ts
@@ -1,0 +1,5 @@
+function doSomething(input: string) {
+	return "doSomething with " + input
+}
+
+export = doSomething;