+++ test/execution-tests/simpleDependency/src/dep.ts
@@ -1,0 +1,3 @@
+import dep = require('./deeperDep');
+
+export = dep;