+++ test/execution-tests/simpleDependency/test/app.tests.ts
@@ -1,0 +1,10 @@
+/// <reference path="../typings/index.d.ts" />
+
+import dep = require('../src/dep');
+
+describe("app", () => {
+  it("dep can be called", () => {
+
+    expect(dep("nothing")).toBe("doSomething with nothing");
+  });
+});